Create passwordMeter.js
Added the Javascript file containing code that checks the input password, looping through all the characters to find if different character variations exists (i.e., uppercase, lowercase, numbers and special characters). Boolean values are used to check what types of characters exist in the input password to be used in the points system later on!