Finishing the Javascript code
The way to test the strength of the given password is to use exponentials. The calculation of how strong the password is by acquiring the different amount of character variations raised by the power of how long the given password is.

Also liked the Javascript file into the html file's script tag. As well as adding the line of code which listens to every input from the user into the password textfield and activates the function strengthChecker()