Update passwordMeter.js
I forgot to add in the boolean variable 'specials' to confirm if a special character(s) exists from the user password input. Also worked on shortening the 'else if' statement for determining special characters by making a string variable 'specialChars' that contains all the special characters from the English keyboard. The new if-statement then uses the current character in the password string in the  for-loop to see if the current iterated character exists within the string 'specialChars'. Thus shortening the if-statement's condition.

Also changed how the program acquires the ascii values of each element by using '.charCodeAt()' function instead of directly acquiring elements from string. This way, the if-statements really do check the ascii values since the function '.charCodeAt()' returns a numeric value instead of an character element.