Added major changes to design of the webpage
I decided to add in colours for how strong/weak a given password is. Red for 'weak', orange for 'Medium' and green for 'STRONG' to help emphasis the strength of the provided password from User.

Also added a 'div' section in the html file consisting of the input password field and the teaxtarea displaying the strength of the password in the input field. The div section can be seen with a light blue box shape area for added style.

Also fixed issues such as 'weak' remaining displayed after user removes all characters in the input password field by adding another if-condition in the function 'strengthChecker()' checking if input length is 0 then display nothing.

Another issue fixed was waiting for a while until a string appears in the tips section. This issue was solved by adding a default tips in the span tag in the html file.