The colour of the border now changes based on the strength of password
I thought about having the colour of the border to change according to the strength of the password similar to the respective colours for 'weak', 'Medium' and 'STRONG'. The change of colours occurs slowly with a transition value of 1 second.

This feature adds a sense of profession to the style of the website.