Added box-shadow to the display div
I experimented adding box-shadow to the 'inputArea' div where its colour matches the colour of the borderline. This is to add more style to the webpage for professional appearance.

To do this, I made a box-shadow attribute in the CSS file. To keep its colour matching the border colour, I used CSS variables and made Javascript code to change the value of the CSS variable using document.documentElement.style.cssText.
A CSS variable --borderColour was made and declared in a new CSS field :root so that it can be access by Javascript via document.documentElement.style.cssText.