Update passwordMeter.css
Modified the CSS code where the "inputArea" div has margin value of 'auto' to keep it in the centre of the screen! Also gave it a max-width value where the larger the webpage is on screen, its width value is as wide as 350px! Also widened the input text field of the passwards object to allow more strings to be displayed when entering longer passwords.