Removed redundant code from Javascript
I figured that since the colour of the text within the h4 tag "result" uses the same colour as the border of the "inputArea" div. I decided to use CSS and the variable --borderColour to be assigned to the colour of the h4 tag "result". This way there can be less code in Javascript replaced by one line in CSS!

Also came up with new ideas for what tips to include in the webpage and added those in!