Update passwordMeter.js
Initially I thought spaces are invalid characters for passwords, but after checking online with cybersecurity websites QandA, it turns out they're allowed.

Hence I removed the else if condition if a space character is present and added a space character in the string 'specialChars' to let the algorithm know that spaces are counted as special characters!