Update passwordMeter.js
Implemented the feature of displaying if password as a variety of character types or not (i.e., upper/lowercase). If the input password only has one character type present, then it will be shown in the weakness section, but if the input password contain more than 2 types of characters, then the strength section will display to User that there exists a good variety of character types.