Update passwordMeter.js
Fixed an issue where if a string consisting of neighbouring characters and is longer than 3 characters in length, the programs keeps incrementing the variable "repeated" for every substring of 3 neighbouring characters found within the string containing neighbouring characters.

This issue is solved by having a while loop which increments the "index" variable until the end of the string containing neighbouring characters.